requirements.txt
# Core dependencies
python-telegram-bot==21.10
openai==1.59.5
python-dotenv==1.0.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
PyGithub==2.1.1

# Advanced features
tavily-python==0.3.0  # Web search
chromadb==0.4.22  # Vector database для памяти
Pillow==10.1.0  # Обработка изображений
requests==2.31.0  # HTTP запросы
instagrapi==2.0.0  # Instagram API
